refactor : 로그인 event module 리팩토링
