+++ lib/presentation/pages/sign_in/sign_in_event.dart
@@ -9,8 +9,10 @@ import 'package:techtalk/presentation/providers/user/user_auth_provider.dart';
 import 'package:techtalk/presentation/providers/user/user_info_provider.dart';
 
 mixin class SignInEvent {
+  ///
   /// 유저 데이터 존재 여부에 따라 라우팅을 분기한다.
-  Future<void> _routeByUserData(WidgetRef ref,
+  ///
+  Future<void> _routeByUserState(WidgetRef ref,
       {required UserAccountProvider accountProvider}) async {
     return ref.watch(userInfoProvider.future).then(
       (userData) async {
@@ -27,7 +29,9 @@ mixin class SignInEvent {
     );
   }
 
-  /// OAuth 인증을 실행한다.
+  ///
+  /// OAuth 인증을 실행
+  ///
   Future<void> _signInOAuth(
     WidgetRef ref, {
     required UserAccountProvider provider,
@@ -38,20 +42,18 @@ mixin class SignInEvent {
           return ref.read(userAuthProvider.notifier).signInOAuth(provider);
         },
       ).then(
-        (_) => _routeByUserData(ref, accountProvider: provider),
+        (_) => _routeByUserState(ref, accountProvider: provider),
       );
     } finally {
       await EasyLoading.dismiss();
     }
   }
 
-  /// 구글 로그인을 눌렀을 때 실행할 콜백
-  Future<void> onTapSignInWithGoogle(WidgetRef ref) async {
-    await _signInOAuth(ref, provider: UserAccountProvider.google);
-  }
-
-  /// 애플 로그인을 눌렀을 때 실행할 콜백
-  Future<void> onTapSignInWithApple(WidgetRef ref) async {
-    await _signInOAuth(ref, provider: UserAccountProvider.apple);
+  ///
+  /// 소셜 로그인 버튼을 클릭 했을 때
+  ///
+  Future<void> onSocialSignInBtnTapped(WidgetRef ref,
+      {required UserAccountProvider socialAccountProvider}) async {
+    await _signInOAuth(ref, provider: socialAccountProvider);
   }
 }
